sửa nhanh UI & chuyển tệp
sửa giao diện YourActivity và di chuyển 2 HighlightModal từ SeenStory sang SeenStoryOwner
