+++ src/Screens/SeenStoryOwner/index.tsx
@@ -10,8 +10,19 @@ import {
   TextInput,
 } from 'react-native';
 import Video from 'react-native-video';
+import { Modalize } from "react-native-modalize";
+import {Portal} from 'react-native-portalize';
 import ModelPeopleSeen from './component/ModelPeopleSeen';
 import ModelSeeMore from './component/ModelSeeMore';
+import HighlightAddModal from './component/HighlightAddModal'
+import HighlightViewModal from './component/HighlightViewModal'
+
+// data mẫu cho modal highlight
+const highlights= [
+  { id: "1", name: "Trip", isAdded: true, imageURL: 'https://i.pinimg.com/736x/5a/92/e7/5a92e7f5a37dbcf79c6740dea218ea52.jpg'},
+  { id: "2", name: "Food", isAdded: true, imageURL: 'https://i.pinimg.com/736x/5a/92/e7/5a92e7f5a37dbcf79c6740dea218ea52.jpg'},
+  { id: "3", name: "Friends", isAdded: false, imageURL: 'https://i.pinimg.com/736x/5a/92/e7/5a92e7f5a37dbcf79c6740dea218ea52.jpg'},
+];
 
 export const SeenStoryOwner = ({route, navigation}: any) => {
   const {selectedItem} = route.params;
@@ -21,6 +32,22 @@ export const SeenStoryOwner = ({route, navigation}: any) => {
   const progressAnim = useRef(new Animated.Value(0)).current;
   const animationRef = useRef(null);
   const videoRef = useRef(null);
+  const viewModalRef = useRef<Modalize>(null);
+  const addModalRef = useRef<Modalize>(null);
+
+  const handleOpenAddModal = () => {
+    viewModalRef.current?.close();
+    setTimeout(() => addModalRef.current?.open(), 300);
+  };
+
+  const handleOnBackAddModal = () => {
+    addModalRef.current?.close();
+    setTimeout(() => viewModalRef.current?.open(), 300);
+  }
+
+  const handleAddHighlight = (name: string) => {
+    // xử lý thêm highlight mới vào danh sách
+  };
 
   const imageDuration = 10000; // 10 seconds for images
 
@@ -170,12 +197,37 @@ export const SeenStoryOwner = ({route, navigation}: any) => {
           />
           <Text style={styles.txtIcon}>Xem thêm</Text>
         </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.viewIconItem}
+          onPress={() => viewModalRef.current?.open()}>
+          <Image
+            style={styles.icon}
+            source={require('../../../assets/icon/heart.png')}
+          />
+        </TouchableOpacity>
         <ModelPeopleSeen visible={visible} onClose={() => setVisible(false)} />
         <ModelSeeMore
           visible={visibleSeeMore}
           onClose={() => setVisibleSeeMore(false)}
         />
       </View>
+
+      <Portal>
+        <HighlightViewModal
+          ref={viewModalRef}
+          data={highlights}
+          onAddNew={handleOpenAddModal}
+        />
+      </Portal>
+            
+      <Portal>
+        <HighlightAddModal
+          ref={addModalRef}
+          onAdd={handleAddHighlight}
+          onBack={handleOnBackAddModal}
+          imageSource={"https://i.pinimg.com/736x/5a/92/e7/5a92e7f5a37dbcf79c6740dea218ea52.jpg"}
+        />
+      </Portal>
     </SafeAreaView>
   );
 };
